% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_plot.R
\name{dendro_plot}
\alias{dendro_plot}
\title{Make a clean dendrogram plot}
\usage{
dendro_plot(dend, rotate = FALSE, x.expand)
}
\arguments{
\item{dend}{a dendrograme, generated by dendrogram_basic function.}

\item{rotate}{logic value. If true, the output plot will be rotated 90 degree
clockwise. Useful to make row wise dendrogram}

\item{x.expand}{numeric. The x expand value. This value is used to make sure
the width of the dendrogram is exactly the same as the heatmap.}
}
\description{
This function takes a dendrogram and make plot using the
\code{\link{ggdendrogram}} function. The returned plot is a ggplot object

This function should be only called from internal
}
\keyword{internal}
