% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram_basic.R
\name{dendrogram_basic}
\alias{dendrogram_basic}
\title{Make a basic dendrogram}
\usage{
dendrogram_basic(data, row.wise = TRUE, seriate = "OLO")
}
\arguments{
\item{data}{the data.}

\item{row.wise}{logic value. If true the row wise dendrogram will be returned.}

\item{seriate}{character. Seriate method to use. "OLO" seems to find the
order of samples (rows or columns) that makes the heatmap with a more clear
pattern. "HC" is the regular hierarchical clustring.}
}
\description{
This function makes a basic dendrogram using the seriation
methods provided by the seriation package

This function should be only called from internal.
}
\keyword{internal}
