% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sider_barplot.R
\name{get_legend}
\alias{get_legend}
\title{Extract the legend from a ggplot}
\usage{
get_legend(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplto object}
}
\description{
This function extracts the legend out of a ggplot object. It is
used to separate the legend from the sider_barplot in the zheatmap funciton.
Should only be called internally.
}
\keyword{internal}
