% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colorkey.R
\name{plot_colorkey}
\alias{plot_colorkey}
\title{Make a color key plot}
\usage{
plot_colorkey(color.range, colors, text.size, show.legend.scale)
}
\arguments{
\item{color.range}{numeric vector that indicates the range of x axis for the
color key}

\item{colors}{character indicates the colors to use}

\item{text.size}{the size of the x axis labels.}

\item{show.legend.scale}{logic value whether to show scales on color key.}
}
\description{
This function makes a color key plot for the heatmap. Should
only be called internally
}
\keyword{internal}
