% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sider_barplot.R
\name{side_barplot}
\alias{side_barplot}
\title{Make an annotation side bar plot}
\usage{
side_barplot(x, id, vertical = F, legend.text.size = 9)
}
\arguments{
\item{x}{a character vector for annotation. It should have the same order
as the input data in the very begining.}

\item{legend.text.size}{numeric indicates the legend text size.}

\item{col.id}{a integer vector to reorder the x vector so it matches the
heatmap.}
}
\description{
This function makes a barplot to put on the side of the heatmap
as annotation. Should only be called internally
}
\keyword{internal}
