% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{zheatmap}
\alias{zheatmap}
\title{Clustered Heatmap}
\usage{
zheatmap(data, colSideBar = NULL, rowSideBar = NULL, Rowv = TRUE,
  Colv = TRUE, colors = viridis(n = 256, alpha = 1, begin = 0, end = 1,
  option = "viridis"), seriate = "OLO", scale = c("column", "row", "none"),
  scale.fun = c("scale", "absolute_scale"), xtext = FALSE, ytext = TRUE,
  text.size = 12, xtext.angle = 0, ytext.angle = 0, xtext.hjust = 1,
  xtext.vjust = 0, ytext.hjust = 1, ytext.vjust = 0,
  legend.text.size = 9, show.legend.scale = FALSE, heights, widths,
  print = TRUE)
}
\arguments{
\item{data}{a data.frame object}

\item{colSideBar}{a character or factor vector to annotate on the column
side between the heatmap ans column-wize dendrogram}

\item{Rowv}{a logical value whether to show the row-wise dendrogram}

\item{Colv}{a logical value whether to show the column-wise dendrogram}

\item{colors}{a character vector off colors to use. The default uses
\code{\link{viridis}}(n=256, alpha = 1, begin = 0, end = 1,
option = "viridis")}

\item{seriate}{character indicates the method to use to calculate the dist.
Usful methods include "OLO", "GW", and "HC". See \code{\link{seriate}} for
more detail}

\item{scale}{A character value indicates whether the data should be scaled
column wise or row wise, or not scaled.}

\item{scale.fun}{the scale function to use. When using "scale" the
\code{\link{scale}} from base package will be used. When using
"absolute_scale" each row or column will be forced between -1 and 1.}

\item{xtext}{logic whether to show x axis labels}

\item{ytext}{logic whether to show y axis labels}

\item{text.size}{numeric}

\item{xtext.angle}{numeric}

\item{ytext.angle}{numeric}

\item{xtext.hjust}{numeric}

\item{xtext.vjust}{numeric}

\item{ytext.hjust}{numeric}

\item{ytext.vjust}{numeric}

\item{show.legend.scale}{logic value whether to show scales on color key. Default is TRUE.}

\item{heights}{numeric vector with length of 2}

\item{print}{logic, whether to print the heatmap out or not. Default is TRUE and if FALSE, a grob object will be returned}

\item{width}{numeric vector with length of 2}
}
\description{
A flexible heatmap function with dendrograms on both sides.
}
\examples{
# using the mtcars data
zheatmap(mtcars)

# show x axis texts and rotate them 90 degree
zheatmap(t(mtcars), scale = "row", xtext.angle = 90)

# change the color
library(RColorBrewer)
my_colors = colorRampPalette(rev(brewer.pal(11, "RdBu")))(256)
zheatmap(mtcars, colors = my_colors)

# use hierachical clustring method
zheatmap(mtcars, seriate = "HC")

# hide the column wise dendrogram
zheatmap(mtcars, Colv = F)

# use color side bar to annotate
zheatmap(mtcars[,-2], rowSideBar = factor(mtcars$cyl))

}
\author{
Chenghao Zhu
}
