% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{zheatmap}
\alias{zheatmap}
\title{Clustered Heatmap}
\usage{
zheatmap(data, colSideBar = NULL, Rowv = TRUE, Colv = TRUE,
  colors = viridis(n = 256, alpha = 1, begin = 0, end = 1, option =
  "viridis"), seriate = "OLO", scale = c("column", "row", "none"),
  scale.fun = c("scale", "absolute_scale"), xtext = FALSE, ytext = TRUE,
  text.size = 12, xtext.angle = 0, ytext.angle = 0, xtext.hjust = 0,
  xtext.vjust = 0, ytext.hjust = 0, ytext.vjust = 0,
  legend.text.size = 11, heights, widths, print = TRUE)
}
\arguments{
\item{data}{a data.frame object}

\item{colSideBar}{a character or factor vector to annotate on the column
side between the heatmap ans column-wize dendrogram}

\item{Rowv}{a logical value whether to show the row-wise dendrogram}

\item{Colv}{a logical value whether to show the column-wise dendrogram}

\item{colors}{a character vector off colors to use. The default uses
\code{\link{viridis}}(n=256, alpha = 1, begin = 0, end = 1,
option = "viridis")}

\item{seriate}{character. See \code{\link{seriate}}}

\item{scale.fun}{the scale function to use. When using "scale" the
\code{\link{scale}} from base package will be used. When using
"absolute_scale" each row or column will be forced between -1 and 1.}

\item{xtext}{logic whether to show x axis labels}

\item{ytext}{logic whether to show y axis labels}

\item{text.size}{numeric}

\item{xtext.angle}{numeric}

\item{xtext.hjust}{numeric}

\item{xtext.vjust}{numeric}

\item{ytext.hjust}{numeric}

\item{ytext.vjust}{numeric}

\item{heights}{numeric vector with length of 2}

\item{print}{logic, whether to print the heatmap out or not. Default is TRUE and if FALSE, a grob object will be returned}

\item{scaled}{A character value indicates whether the data should be scaled
column wise or row wise, or not scaled.}

\item{xtext.angle}{numeric}

\item{width}{numeric vector with length of 2}
}
\description{
A flexible heatmap function with dendrograms on both sides.
}
\author{
Chenghao Zhu
}
